minor file refactor
