+++ src/main/java/com/mycmsbackend/repository/RefreshTokenRepository.java
@@ -1,7 +1,7 @@
 package com.mycmsbackend.repository;
 
-import com.mycmsbackend.model.RefreshToken;
-import com.mycmsbackend.model.User;
+import com.mycmsbackend.domain.RefreshToken;
+import com.mycmsbackend.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
