+++ src/main/java/com/mycmsbackend/security/advice/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.mycmsbackend.advice;
+package com.mycmsbackend.security.advice;
 
 import java.util.Date;
