--- src/main/java/com/mycmsbackend/service/AccountsService.java
@@ -1,23 +1,0 @@
-package com.mycmsbackend.service;//package com.mycmsbackend.service;
-//
-//import com.mycmsbackend.model.Accounts;
-//import com.mycmsbackend.repository.AccountsRepository;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.stereotype.Service;
-//
-//import java.util.List;
-//
-//@Service("accountsService")
-//public class AccountsService {
-//
-//    private final AccountsRepository accountsRepository;
-//
-//    @Autowired
-//    public AccountsService(AccountsRepository accountsRepository) {
-//        this.accountsRepository = accountsRepository;
-//    }
-//
-//    public List<Accounts> getAllAccounts(){
-//        return accountsRepository.findAll();
-//    }
-//}
