+++ src/main/java/com/mycmsbackend/service/UsersService.java
@@ -1,6 +1,6 @@
 package com.mycmsbackend.service;
 
-import com.mycmsbackend.model.User;
+import com.mycmsbackend.domain.User;
 import com.mycmsbackend.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
