+++ src/main/java/com/mycmsbackend/service/dto/OrderDTO.java
@@ -1,0 +1,25 @@
+//package com.mycmsbackend.domain.dto;
+//
+//import lombok.AllArgsConstructor;
+//import lombok.Data;
+//import lombok.NoArgsConstructor;
+//
+///**
+// * @Data is a convenient shortcut annotation that bundles the features of @ToString, @EqualsAndHashCode, @Getter / @Setter and @RequiredArgsConstructor together:
+// * In other words, @Data generates all the boilerplate that is normally associated with simple POJOs (Plain Old Java Objects)
+// * and beans: getters for all fields, setters for all non-final fields, and appropriate toString,
+// * equals and hashCode implementations that involve the fields of the class, and a constructor that initializes all final fields, as well as all non-final fields with no initializer that have been marked with @NonNull, in order to ensure the field is never null.
+// */
+//
+//@Data
+//@AllArgsConstructor
+//@NoArgsConstructor
+//public class OrderDTO {
+//
+//    private String orderId;
+//    private String orderName;
+//    private String orderEmailRecipient;
+//    private int orderQty;
+//    private double orderPrice;
+//
+//}
