+++ src/main/java/com/mycmsbackend/service/dto/OrderEventDTO.java
@@ -1,0 +1,20 @@
+//package com.mycmsbackend.domain.dto;
+//
+//import lombok.AllArgsConstructor;
+//import lombok.Data;
+//import lombok.NoArgsConstructor;
+//
+///**
+// * This class going to be used for transfer data between Producer and Consumer using apache kafka
+// */
+//
+//@Data
+//@AllArgsConstructor
+//@NoArgsConstructor
+//public class OrderEventDTO {
+//
+//    private String message;
+//    private String status;
+//    private OrderDTO orderDto;
+//
+//}
